+++ Utilities/Mines.cs
@@ -41,7 +41,7 @@ namespace SeedFinding
             return CheckStone1_5(x*1000 + y + floor + gameId / 2, ladder,geologist,excavator,floor);
         }
 
-        public static List<int> CheckStone1_6(int day, int gameId, int floor, int x, int y, bool ladder = false, bool geologist = false, bool excavator = false)
+        public static List<int> CheckStone1_6(int day, uint gameId, int floor, int x, int y, bool ladder = false, bool geologist = false, bool excavator = false)
         {
 
             Random r = Utility.CreateDaySaveRandom(day, gameId, x * 1000, y, floor);
@@ -111,9 +111,9 @@ namespace SeedFinding
             return value;
         }
 
-        public static (string,int) GetGeodeContents1_6(int gameId, int geodesCracked, Geode whichGeode, int deepestMineLevel=0, bool qibeans = false, bool farmingMastery=false, int fishingLevel=0, bool hasCoconutHat = false, bool hasMysteryBook = false)
+        public static (string,int) GetGeodeContents1_6(uint gameId, int geodesCracked, Geode whichGeode, int deepestMineLevel=0, bool qibeans = false, bool farmingMastery=false, int fishingLevel=0, bool hasCoconutHat = false, bool hasMysteryBook = false)
         {
-            Random r = Utility.CreateRandom(geodesCracked, gameId / 2); 
+            Random r = Utility.CreateRandom(geodesCracked, gameId / 2uL); 
             int prewarm_amount = r.Next(1, 10);
             for (int i = 0; i < prewarm_amount; i++)
             {
@@ -504,7 +504,7 @@ namespace SeedFinding
                 _ => ("(O)382", amount),
             };
         }
-        public static void PrintGeodeContents(int gameId, int startingGeode, int count, List<Geode> geodeTypes, string delimiter, bool excludeOres=true, int deepestMineLevel=0, bool qibeans = false, bool printBestGeode=false, int printBestGeodeMinPrice=0, bool before1_5=false)
+        public static void PrintGeodeContents(uint gameId, int startingGeode, int count, List<Geode> geodeTypes, string delimiter, bool excludeOres=true, int deepestMineLevel=0, bool qibeans = false, bool printBestGeode=false, int printBestGeodeMinPrice=0, bool before1_5=false)
         {
             List<string> unsellables = new List<string>() { "100", "101", "103", "104", "105", "106", "108", "109", "110", "111", "112", "113", "114", "115", "116", "117", "118", "119", "120", "121", "122", "123", "124", "125", "330", "390" };
             if (excludeOres)
@@ -551,9 +551,9 @@ namespace SeedFinding
 
         
 
-        public static (int,int) GetGeodeContents(int gameId, int geodesCracked, Geode whichGeode, int deepestMineLevel=0, bool qibeans=false, bool before1_5=false)
+        public static (int,int) GetGeodeContents(uint gameId, int geodesCracked, Geode whichGeode, int deepestMineLevel=0, bool qibeans=false, bool before1_5=false)
         {
-            Random r = new Random(geodesCracked + gameId / 2);
+            Random r = new Random(geodesCracked + (int)gameId / 2);
             int prewarm_amount;
             prewarm_amount = r.Next(1, 10);
             for (int j = 0; j < prewarm_amount; j++)
